usart support in a small driver lib
